+++ packages/form/src/index.tsx
@@ -46,6 +46,7 @@ import BetaSchemaForm from './components/SchemaForm';
 import './index.less';
 import { ProFormContext } from '@ant-design/pro-utils';
 import type { LightFilterFooterRender, ProFormFieldItemProps } from './interface';
+import { FormItemProvide } from './components/FormItem';
 
 export type {
   ProFormFieldProps,
@@ -85,6 +86,7 @@ export {
   ProFormField,
   ProFormList,
   ProFormSelect,
+  FormItemProvide,
   ProFormDigit,
   ProFormMoney,
   ProFormCheckbox,