+++ packages/table/src/Table.tsx
@@ -77,6 +77,7 @@ function TableRender<T extends Record<string, any>, U, ValueType>(
     style,
     cardProps,
     alertDom,
+    name,
     onSortChange,
     onFilterChange,
     options,
@@ -131,7 +132,6 @@ function TableRender<T extends Record<string, any>, U, ValueType>(
    */
   const editableDataSource = (): T[] => {
     const { options: newLineOptions, defaultValue: row } = editableUtils.newLineRecord || {};
-
     if (newLineOptions?.position === 'top') {
       return [row, ...action.dataSource];
     }
@@ -211,27 +211,41 @@ function TableRender<T extends Record<string, any>, U, ValueType>(
       )
     : baseTableDom;
 
-  const tableContentDom = (
-    <>
-      {toolbarDom}
-      {alertDom}
-      {props.editable ? (
-        <ProForm
-          {...props.editable?.formProps}
-          component={false}
-          form={props.editable?.form}
-          onValuesChange={editableUtils.onValuesChange}
-          key="table"
-          submitter={false}
-          omitNil={false}
-        >
-          {tableDom}
-        </ProForm>
-      ) : (
-        tableDom
-      )}
-    </>
-  );
+  useEffect(() => {
+    counter.setPrefixName(name);
+  }, [name]);
+
+  const tableContentDom = useMemo(() => {
+    if (props.editable) {
+      return (
+        <>
+          {toolbarDom}
+          {alertDom}
+          <ProForm
+            {...props.editable?.formProps}
+            component={false}
+            form={props.editable?.form}
+            onValuesChange={editableUtils.onValuesChange}
+            key="table"
+            submitter={false}
+            omitNil={false}
+          >
+            {tableDom}
+          </ProForm>
+        </>
+      );
+    }
+
+    return (
+      <>
+        {toolbarDom}
+        {alertDom}
+        {tableDom}
+      </>
+    );
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [alertDom, !!props.editable, tableDom, toolbarDom]);
+
   /** Table 区域的 dom，为了方便 render */
   const tableAreaDom =
     cardProps === false ? (
@@ -328,6 +342,7 @@ const ProTable = <T extends Record<string, any>, U extends ParamsType, ValueType
     onColumnsStateChange,
     options,
     search,
+    name: isEditorTable,
     onLoadingChange,
     rowSelection: propsRowSelection = false,
     beforeSearchSubmit,
@@ -522,6 +537,7 @@ const ProTable = <T extends Record<string, any>, U extends ParamsType, ValueType
   /** 可编辑行的相关配置 */
   const editableUtils = useEditableArray<any>({
     ...props.editable,
+    tableName: props.name,
     getRowKey,
     childrenColumnName: props.expandable?.childrenColumnName,
     dataSource: action.dataSource || [],
@@ -585,7 +601,8 @@ const ProTable = <T extends Record<string, any>, U extends ParamsType, ValueType
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [
     propsColumns,
-    counter,
+    counter?.sortKeyColumns,
+    counter?.columnsMap,
     columnEmptyText,
     type,
     // eslint-disable-next-line react-hooks/exhaustive-deps
@@ -691,6 +708,7 @@ const ProTable = <T extends Record<string, any>, U extends ParamsType, ValueType
   return (
     <TableRender
       {...props}
+      name={isEditorTable}
       rootRef={rootRef}
       size={counter.tableSize}
       onSizeChange={counter.setTableSize}